+++ joblib/test/test_disk.py
@@ -2,9 +2,12 @@
 Unit tests for the disk utilities.
 """
 
-# Author: Gael Varoquaux <gael dot varoquaux at normalesup dot org>
+# Authors: Gael Varoquaux <gael dot varoquaux at normalesup dot org>
+#          Lars Buitinck <L.J.Buitinck@uva.nl>
 # Copyright (c) 2010 Gael Varoquaux
 # License: BSD Style, 3 clauses.
+
+from __future__ import with_statement
 
 import os
 import shutil
@@ -13,7 +16,7 @@ from tempfile import mkdtemp
 
 import nose
 
-from ..disk import memstr_to_kbytes, disk_used
+from ..disk import disk_used, memstr_to_kbytes, mkdirp
 
 
 ###############################################################################
@@ -30,9 +33,10 @@ def test_disk_used():
         a = array.array('i')
         sizeof_i = a.itemsize
         target_size = 1024
-        n = target_size * 1024 / sizeof_i
+        n = int(target_size * 1024 / sizeof_i)
         a = array.array('i', n * (1,))
-        a.tofile(file(os.path.join(cachedir, 'test'), 'w'))
+        with open(os.path.join(cachedir, 'test'), 'wb') as output:
+            a.tofile(output)
         nose.tools.assert_true(disk_used(cachedir) >= target_size)
         nose.tools.assert_true(disk_used(cachedir) < target_size + 12)
     finally:
@@ -41,7 +45,22 @@ def test_disk_used():
 
 def test_memstr_to_kbytes():
     for text, value in zip(('80G', '1.4M', '120M', '53K'),
-                           (80 * 1024**2, int(1.4 * 1024), 120 * 1024, 53)):
+                           (80 * 1024 ** 2, int(1.4 * 1024), 120 * 1024, 53)):
         yield nose.tools.assert_equal, memstr_to_kbytes(text), value
 
     nose.tools.assert_raises(ValueError, memstr_to_kbytes, 'foobar')
+
+
+def test_mkdirp():
+    try:
+        tmp = mkdtemp()
+
+        mkdirp(os.path.join(tmp, "ham"))
+        mkdirp(os.path.join(tmp, "ham"))
+        mkdirp(os.path.join(tmp, "spam", "spam"))
+
+        # Not all OSErrors are ignored
+        nose.tools.assert_raises(OSError, mkdirp, "")
+
+    finally:
+        shutil.rmtree(tmp)