+++ joblib/test/test_parallel.py
@@ -6,12 +6,19 @@ Test the parallel module.
 # Copyright (c) 2010-2011 Gael Varoquaux
 # License: BSD Style, 3 clauses.
 
-import time
+import time, sys
 try:
     import cPickle as pickle
     PickleError = TypeError
 except:
     import pickle
+    PickleError = pickle.PicklingError
+try:
+    from io import BytesIO
+except:
+    from cStringIO import StringIO as BytesIO
+
+if sys.version_info[0] == 3:
     PickleError = pickle.PicklingError
 
 from ..parallel import Parallel, delayed, SafeFunction, WorkerInterrupt, \
@@ -28,7 +35,7 @@ def division(x, y):
 
 
 def square(x):
-    return x**2
+    return x ** 2
 
 
 def exception_raiser(x):
@@ -46,7 +53,7 @@ def f(x, y=0, z=0):
     """ A module-level function so that it can be spawn with
     multiprocessing.
     """
-    return x**2 + y + z
+    return x ** 2 + y + z
 
 
 ###############################################################################
@@ -57,10 +64,39 @@ def test_cpu_count():
 ###############################################################################
 # Test parallel
 def test_simple_parallel():
-    X = range(10)
-    for n_jobs in (1, 2, -1):
+    X = range(5)
+    for n_jobs in (1, 2, -1, -2):
         yield (nose.tools.assert_equal, [square(x) for x in X],
-                        Parallel(n_jobs=-1)(delayed(square)(x) for x in X))
+                Parallel(n_jobs=-1)(
+                        delayed(square)(x) for x in X))
+    try:
+        # To smoke-test verbosity, we capture stdout
+        orig_stdout = sys.stdout
+        sys.stdout = BytesIO()
+        orig_stderr = sys.stdout
+        sys.stderr = BytesIO()
+        for verbose in (2, 11, 100):
+                Parallel(n_jobs=-1, verbose=verbose)(
+                        delayed(square)(x) for x in X)
+                Parallel(n_jobs=1, verbose=verbose)(
+                        delayed(square)(x) for x in X)
+                Parallel(n_jobs=2, verbose=verbose, pre_dispatch=2)(
+                        delayed(square)(x) for x in X)
+    except Exception, e:
+        print sys.stdout.getvalue()
+        print sys.stderr.getvalue()
+        raise e
+    finally:
+        sys.stdout = orig_stdout
+        sys.stderr = orig_stderr
+
+
+def nested_loop():
+    Parallel(n_jobs=2)(delayed(square)(.01) for _ in range(2))
+
+
+def test_nested_loop():
+    Parallel(n_jobs=2)(delayed(nested_loop)() for _ in range(2))
 
 
 def test_parallel_kwargs():
@@ -79,7 +115,7 @@ def test_parallel_pickling():
         that cannot be pickled.
     """
     def g(x):
-        return x**2
+        return x ** 2
     nose.tools.assert_raises(PickleError,
                              Parallel(),
                              (delayed(g)(x) for x in range(10))
@@ -111,11 +147,12 @@ def test_error_capture():
                     [delayed(division)(x, y) for x, y in zip((0, 1), (1, 0))],
                         )
     try:
+        ex = JoblibException
         Parallel(n_jobs=1)(
                     delayed(division)(x, y) for x, y in zip((0, 1), (1, 0)))
-    except Exception, e:
-        pass
-    nose.tools.assert_false(isinstance(e, JoblibException))
+    except Exception as e:
+        ex = e
+    nose.tools.assert_false(isinstance(ex, JoblibException))
 
 
 class Counter(object):